Move apollo logic into one file
+ Use http-link from http boost
